make a pass through the documentation
